Add function to get the choice of the user
Created function getUserChoice. It asks the user to enter Rock, Paper or Scissors in a prompt.
Then it transforms the input value into lowercase (in case the user spelled it with an uppercase letter).
It also checks if the user choice equals one of the needed terms and if not it logs a message to the console.