Add a play game function to play 5 rounds
To do this I had to rework the functions, so they are called again when playRound is called.
The game is also comparing how many rounds each player has won and declares a final winner.

There is still a bug when the input doesn't equals one of the terms.
Then the returned value is is the first (wrong) inputted value and not the new one.